Moved main content to its own file. Removed some route protection
