+++ frontend/src/App.js
@@ -7,57 +7,30 @@ import {
   Navigate,
 } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import HomePage from './pages/Home';
-import NotFound from './pages/NotFound';
-import Navigation from './components/Navigation';
 import LoginPage from './pages/LoginPage';
 import SignupPage from './pages/SignupPage';
 import { UserContext } from './auth/userContext';
+import MainContent from './components/MainContent';
 
 function App() {
-  const { isAuthenticated } = useSelector((state) => state.auth);
-
   return (
     <UserContext>
       <BrowserRouter>
         <Routes>
           <Route element={<AnonymousRoute />}>
-            {/* These routes only available to those NOT logged in */}
-            <Route
-              exact
-              path='/login'
-              element={
-                isAuthenticated ? <Navigate to='/' replace /> : <LoginPage />
-              }
-            />
-            <Route
-              exact
-              path='/signup'
-              element={
-                isAuthenticated ? <Navigate to='/' replace /> : <SignupPage />
-              }
-            />
+            {/* 
+              These routes only available to those NOT logged in.
+              This just means that if you are signed in then you cannot see the login or signup page.
+            */}
+            <Route exact path='/login' element={<LoginPage />}/>
+            <Route exact path='/signup' element={<SignupPage />}/>
           </Route>
-          <Route
-            path='*'
-            element={isAuthenticated ? <MainContent /> : <LoginPage />}
-          />
+          <Route path='*'element={<MainContent />} />
         </Routes>
       </BrowserRouter>
     </UserContext>
   );
 }
-
-const MainContent = () => {
-  return (
-    <Navigation>
-      <Routes>
-        <Route path='/' element={<HomePage />} />
-        <Route path='*' element={<NotFound />} />
-      </Routes>
-    </Navigation>
-  );
-};
 
 const AnonymousRoute = () => {
   const { isAuthenticated } = useSelector((state) => state.auth);
