--- frontend/src/components/Navigation.jsx
@@ -1,36 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import NavBar from './NavBar';
-
-const Navigation = ({ children }) => {
-  return (
-    <div style={{ overflow: 'hidden', height: '100vh' }}>
-      <NavBar />
-      <div
-        style={{
-          overflow: 'scroll',
-          backgroundImage: 'url(/background.jpeg)',
-          backgroundSize: 'cover',
-          backgroundPosition: 'center',
-        }}
-        className='container-fluid h-100'
-      >
-        {/* <div className='row h-100'> */}
-        {/*   <div className='col-2' id='green' style={{ maxWidth: '150px' }}> */}
-        {/*     <SideBar /> */}
-        {/*   </div> */}
-        {/*   <div className='col' style={{ textAlign: 'justify', backgroundColor: 'grey' }}> */}
-        {/*     {children} */}
-        {/*   </div> */}
-        {/* </div> */}
-        {children}
-      </div>
-    </div>
-  );
-};
-
-Navigation.propTypes = {
-  children: PropTypes.node,
-};
-
-export default Navigation;
